pyproject.toml
[project]
name = "hobby-off-generator"
version = "0.1.0"
description = "Hobby OFF image generator"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.0",
    "pillow>=10.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.1",
]

# ----------------------------
# Ruff
# ----------------------------
[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["ALL"]

ignore = [
    "D1",
    "D203",
    "D213",
    "D400",
    "D415",
    "TD",
    "T201",
    "TRY003",
    "EM101",
    "COM812",
    "PLR0913",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "S101",
    "PLR2004",
    "ANN",
    "D",
]

# ----------------------------
# Pytest（★追加部分）
# ----------------------------
[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "-ra"

# ----------------------------
# Pyright
# ----------------------------
[tool.pyright]
typeCheckingMode = "strict"
pythonVersion = "3.12"
reportUnusedFunction = false